Fix #98
Both #94/#98 were fixed using the remove + add to collection strategy.
The reason this has to be done is CollectionChanged doesn’t
automatically track changes within an object. To do this, you have to
forward PropertyChanged calls for all properties to the
CollectionChanged event. Kind of a huge pain and creates tons of
overhead. Better to just leave it out for now.